#3239: BytesOf refactoring
